Refactor data source plugin into a datasource package
This provides a cleaner separation between the different plugin types
now that we have multiple types of plugins (since adding provision
plugins).
