+++ cmd/cofidectl/cmd/apbinding/apbinding.go
@@ -10,7 +10,7 @@ import (
 	ap_binding_proto "github.com/cofide/cofide-api-sdk/gen/go/proto/ap_binding/v1alpha1"
 	trust_zone_proto "github.com/cofide/cofide-api-sdk/gen/go/proto/trust_zone/v1alpha1"
 	cmdcontext "github.com/cofide/cofidectl/pkg/cmd/context"
-	cofidectl_plugin "github.com/cofide/cofidectl/pkg/plugin"
+	"github.com/cofide/cofidectl/pkg/plugin/datasource"
 
 	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
@@ -85,7 +85,7 @@ func (c *APBindingCommand) GetListCommand() *cobra.Command {
 	return cmd
 }
 
-func (c *APBindingCommand) list(source cofidectl_plugin.DataSource, opts ListOpts) ([]*ap_binding_proto.APBinding, error) {
+func (c *APBindingCommand) list(source datasource.DataSource, opts ListOpts) ([]*ap_binding_proto.APBinding, error) {
 	var err error
 	var trustZones []*trust_zone_proto.TrustZone
