+++ pkg/plugin/provision/plugin.go
@@ -10,7 +10,6 @@ import (
 
 	cofidectl_proto "github.com/cofide/cofide-api-sdk/gen/go/proto/cofidectl_plugin/v1alpha1"
 	provisionpb "github.com/cofide/cofide-api-sdk/gen/go/proto/provision_plugin/v1alpha1"
-	"github.com/cofide/cofidectl/pkg/plugin"
 	"github.com/cofide/cofidectl/pkg/plugin/datasource"
 	go_plugin "github.com/hashicorp/go-plugin"
 	"google.golang.org/grpc"
@@ -120,7 +119,7 @@ func (c *ProvisionPluginClientGRPC) TearDown(ctx context.Context, source datasou
 // This uses the bidirectional communication feature of go-plugin. See
 // https://pkg.go.dev/github.com/hashicorp/go-plugin/examples/bidirectional for an example.
 func (c *ProvisionPluginClientGRPC) startDataSourceServer(source datasource.DataSource) (*grpc.Server, uint32) {
-	dsServer := &plugin.GRPCServer{Impl: source}
+	dsServer := &datasource.GRPCServer{Impl: source}
 
 	serverCh := make(chan *grpc.Server)
 	serverFunc := func(opts []grpc.ServerOption) *grpc.Server {
@@ -230,6 +229,6 @@ func (s *GRPCServer) getDataSourceClient(ctx context.Context, dataSourceID uint3
 		return nil, nil, err
 	}
 
-	client := plugin.NewDataSourcePluginClientGRPC(ctx, cofidectl_proto.NewDataSourcePluginServiceClient(conn))
+	client := datasource.NewDataSourcePluginClientGRPC(ctx, cofidectl_proto.NewDataSourcePluginServiceClient(conn))
 	return client, conn, nil
 }
